feature: move two context and refresh notify platforms
